requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
pandas
psycopg2-binary
pymongo
pyjwt
python-jose[cryptography] 
passlib[bcrypt]
python-multipart